macro system simplifies md and supports variables
Instead of the complex HTML in the md, I've replaced this with a new system that uses a combination of variables like
$name: JD Lien

and macros like
{{HEADING}}
that significantly simplify the markup of the document to make for easier editing.

resume.py now integrates this support internally in a flexible way.

Also, you can now keep variants in subfolders so you can keep track of many copies of a resume while maintaining any filename you want the recipient to see.